Improve unit testing script

+++ grouper/misc/ci-test/ci-test-hsql.sh
@@ -17,72 +17,88 @@
 
 GROUPER_HOME=/var/grouper-ci/git/grouper
 GROUPER_LOGDIR=/var/grouper-ci/log
-LOGFILE=$GROUPER_LOGDIR/grouper-ci-test-$(date +%Y%m%d_%H%M%S)-$$.log
+LOGTAG=$(date +%Y%m%d_%H%M%S)-$$
+TESTLOG=$GROUPER_LOGDIR/grouper-ci-test-$LOGTAG.log
+SUMMARYLOG=$GROUPER_LOGDIR/grouper-ci-test-$LOGTAG.summary.log
+BUILDLOG=$GROUPER_LOGDIR/grouper-ci-mvn-$LOGTAG.log
+
+# Whether to still run the tests even if there are no git updates
+# can enable these during debugging
+#FORCE_TESTS=true
+#SKIP_PULL=true
 
 MVN=/opt/apache-maven-3.6.3/bin/mvn
 
-echo $(date) "CI test started" > $LOGFILE
-echo '----------------------' >>$LOGFILE
+echo $(date) "CI test started" > $TESTLOG
+echo '----------------------' >>$TESTLOG
 
 if [ ! -d "$GROUPER_HOME" ]; then
-  echo "Grouper home '$GROUPER_HOME' missing -- skipping tests" >>$LOGFILE
+  echo "Grouper home '$GROUPER_HOME' missing -- skipping tests" >>$TESTLOG
   exit 1
 fi
 
 cd $GROUPER_HOME
-git checkout master >>$LOGFILE 2>&1
+if [ "$SKIP_PULL" = "" ]; then
+  echo "Pulling from master" >>$TESTLOG 2>&1
+  git checkout master >>$TESTLOG 2>&1
 
-PULL_OUTFILE=$(mktemp --tmpdir=$GROUPER_LOGDIR)
-git pull > $PULL_OUTFILE
+  PULL_OUTFILE=$(mktemp --tmpdir=$GROUPER_LOGDIR)
+  git pull > $PULL_OUTFILE
 
-echo "Git pull output:" >>$LOGFILE
-cat $PULL_OUTFILE >>$LOGFILE
+  echo "Git pull output:" >>$TESTLOG
+  cat $PULL_OUTFILE >>$TESTLOG
 
-if [ -z $PULL_OUTFILE -o "Already up-to-date." = "$(cat $PULL_OUTFILE)" ]; then
-  echo "Git already up to date -- exiting" >>$LOGFILE
+  if [ -z $PULL_OUTFILE -o "Already up-to-date." = "$(cat $PULL_OUTFILE)" ]; then
+    if [ "$FORCE_TESTS" = "true" ]; then
+      echo "Git already up to date, but continuing since FORCE_TESTS=true" >>$TESTLOG
+    else
+      echo "Git already up to date -- exiting" >>$TESTLOG
+      rm $PULL_OUTFILE
+      exit 0
+    fi
+  fi
   rm $PULL_OUTFILE
+else
+  echo "Skipping Git pull due to SKIP_PULL=$SKIP_PULL" >>$TESTLOG 2>&1
 fi
-exit 0
 
-rm $PULL_OUTFILE
-
-
-# Note, prettier version of emails is --format='%an <%ae>'
-# apt install mailx
-COMMITTER_EMAILS=$(git --no-pager show -s --format=%ae master@{1}..master | sort -u)
-
-if [ -z "$COMMITTER_EMAILS" ]; then
-  echo "No committer emails found for latest pull -- skipping tests" >>$LOGFILE
-  exit 1
-fi
+## Note, prettier version of emails is --format='%an <%ae>'
+## apt install mailx
+#COMMITTER_EMAILS=$(git --no-pager show -s --format=%ae master@{1}..master | sort -u)
+#
+#if [ -z "$COMMITTER_EMAILS" ]; then
+#  echo "No committer emails found for latest pull -- skipping tests" >>$TESTLOG
+#  exit 1
+#fi
+COMMITTER_EMAILS=unused
 
 # Check for '[skip tests]' in the commit text
 # Note: grep -q returns 0 if found and 1 if not, so flag is the opposite of expected
-echo "Checking for [skip tests]" >>$LOGFILE
-git --no-pager show -s --format="%s %b" master@{1}..master 2>>$LOGFILE | grep -q '\[skip tests\]'
+echo "Checking for [skip tests]" >>$TESTLOG
+git --no-pager show -s --format="%s %b" master@{1}..master 2>>$TESTLOG | grep -q '\[skip tests\]'
 exit_code=$?
 if [ $exit_code -eq 0 ]; then
-  echo "Found '[skip tests]' within latest pull -- skipping tests" >>$LOGFILE
+  echo "Found '[skip tests]' within latest pull -- skipping tests" >>$TESTLOG
   exit 1
 fi
 
 # Build (all projects) and copy dependencies for grouper subproject, so all can be run from the target/ directory
 # NOTE: This needs to be java 8, not 11!
-echo "Building Maven projects" >>$LOGFILE
-$MVN -f grouper-parent clean package install  >>$LOGFILE 2>&1
+echo "Building Maven projects (logged to $BUILDLOG)" >>$TESTLOG
+$MVN -f grouper-parent clean package install  >>$BUILDLOG 2>&1
 exit_code=$?
 if [ $exit_code -ne 0 ]; then
-  echo "Maven build failed (exit $exit_code)" >>$LOGFILE
+  echo "Maven build failed (exit $exit_code)" >>$TESTLOG
   exit 1
 fi
 
-echo "Downloading Grouper api dependencies" >>$LOGFILE
+echo "Downloading Grouper api dependencies" >>$TESTLOG
 #since we are testing in this script, we don't want to skip the test dependencies
 #$MVN -f grouper dependency:copy-dependencies -DincludeScope=runtime
-$MVN -f grouper dependency:copy-dependencies >>$LOGFILE 2>&1
+$MVN -f grouper dependency:copy-dependencies >>$BUILDLOG 2>&1
 exit_code=$?
 if [ $exit_code -ne 0 ]; then
-  echo "Maven dependency:copy-dependencies failed (exit $exit_code)" >>$LOGFILE
+  echo "Maven dependency:copy-dependencies failed (exit $exit_code)" >>$TESTLOG
   exit 1
 fi
 
@@ -102,7 +118,7 @@ java -cp grouper/target/dependency/hsqldb-2.3.5.jar org.hsqldb.server.Server \
   --database.0 "mem:grouper;user=grouper;password=test" \
   --dbname.0 grouper \
   --address 127.0.0.1 \
-  --port 9101  >>$LOGFILE 2>&1 &
+  --port 9101  >>$TESTLOG 2>&1 &
 
 
 # Set up the classpath to use the newly created artifacts
@@ -119,13 +135,13 @@ CP=$CP:grouper/conf
 #grouper/conf
 
 # Init the grouper database
-echo "Initializing the Grouper database" >>$LOGFILE
+echo "Initializing the Grouper database" >>$TESTLOG
 chmod u+x grouper/bin/gsh.sh
-CLASSPATH="$CP" GROUPER_CONF="grouper/misc/ci-test/confForTestHSQL" grouper/bin/gsh.sh -registry -runscript -noprompt >>$LOGFILE 2>&1
+CLASSPATH="$CP" GROUPER_CONF="grouper/misc/ci-test/confForTestHSQL" grouper/bin/gsh.sh -registry -runscript -noprompt >>$TESTLOG 2>&1
 
 exit_code=$?
 if [ $exit_code -ne 0 ]; then
-  echo "Failed to init the database (exit $exit_code)" >>$LOGFILE
+  echo "Failed to init the database (exit $exit_code)" >>$TESTLOG
   exit 1
 fi
 
@@ -133,17 +149,17 @@ fi
 # Run the tests
 
 # make sure this runs with Java 8
-which java >>$LOGFILE
+which java >>$TESTLOG
 JAVA=java
 
-for var in PWD COMMITTER_EMAILS LOGFILE CP; do
-  echo "$var = ${!var}" >>$LOGFILE
+for var in PWD COMMITTER_EMAILS TESTLOG CP; do
+  echo "$var = ${!var}" >>$TESTLOG
 done
 
-echo $(date) " | START" >>$LOGFILE
+echo $(date) " | START" >>$TESTLOG
 
 # clean out logs from previous run
-rm -f grouper/logs/*.log >>$LOGFILE 2>&1
+rm -f grouper/logs/*.log >>$TESTLOG 2>&1
 
 echo "Executing edu.internet2.middleware.grouper.AllTests"
 $JAVA -classpath "$CP" \
@@ -152,11 +168,11 @@ $JAVA -classpath "$CP" \
   -XX:MaxPermSize=300m -Xms80m -Xmx640m \
   edu.internet2.middleware.grouper.AllTests \
   -all -noprompt \
-  >>$LOGFILE 2>&1
+  >>$TESTLOG 2>&1
 
 exit_code=$?
 
-echo $(date) "CI test started (exit code $exit_code)" >>$LOGFILE
+echo $(date) "CI test started (exit code $exit_code)" >>$TESTLOG
 
 
 GROUPER_ATTACH=
@@ -164,13 +180,23 @@ if [ -n grouper/logs/grouper_error.log]; then
   GROUPER_ATTACH="-a grouper/logs/grouper_error.log"
 fi
 
-jobs >>$LOGFILE
-kill %1 >>$LOGFILE 2>&1
+jobs >>$TESTLOG
+kill %1 >>$TESTLOG 2>&1
 
 #DEBUG
 COMMITTER_EMAILS=chad_redman@unc.edu
 
-echo "CI completed tests" | mailx -s "CI test results" -a $LOGFILE $GROUPER_ATTACH $COMMITTER_EMAILS 2>&1
+tail -n +$(( $(egrep -n '^Time: ' $TESTLOG | tail -n1 | cut -d: -f1) )) $TESTLOG >$SUMMARYLOG 2>>$TESTLOG
+summary_code=$?
+if [ $summary_code -ne 0 -o -z $SUMMARYLOG ]; then
+  echo "DEBUG: summary_code=$summary_code" >>$TESTLOG 2>&1
+  ls -alFd $SUMMARYLOG >>$TESTLOG 2>&1
+  echo "CI completed tests (summary failed)" | mailx -s "CI test results" -a $TESTLOG -a $BUILDLOG $GROUPER_ATTACH $COMMITTER_EMAILS 2>&1
+  exit $exit_code
+fi
+
+echo "CI completed tests" | mailx -s "CI test results (summary)" -a $SUMMARYLOG $COMMITTER_EMAILS 2>&1
 
 # Exit with the result from the test run
 exit $exit_code
+
